Widen index block start offset to long; partitions past 2GiB corrupted the index
CursorIndexWriter.indexBlockStartOffset was an int, unlike the reference
SortedTablePartitionWriter's long. Partitions larger than 2GiB wrap the offset negative past
Integer.MAX_VALUE, and the block-cut arithmetic then cuts every row — corrupting the
serialized offsets in the promoted index. Pinned by the giant-partition boundary run
(ck_stride = rows_per_sstable, forcing disjoint windows so the merged partition size is
maximized rather than halved by overlap).

Also: the differential harness's structural verifier debug-logs every index block during
its extended walk (~560K lines for a >2GiB partition), and ant's junit formatter buffers all
test output in memory — at that log volume the buffering, not the verification itself, OOMs
the test fork. The verifier's debug stream is now silenced in scale mode.

Also corrects the giant-partition scenario's own math: the logged/documented partition size
used sstables * rows_per_sstable (the total PRE-MERGE input rows), not the actual MERGED
partition size, which with the default half-window overlap is (sstables-1) * ck_stride +
rows_per_sstable — a smaller number. A prior "2.6GiB" boundary run under this miscount
actually reached only ~1.4GiB. ck_stride is now a knob (defaulting to the existing
half-window overlap); setting it equal to rows_per_sstable produces disjoint windows and the
true worst-case merged partition size, needed to genuinely exercise the 2GiB boundary.

+++ src/java/org/apache/cassandra/io/sstable/CursorIndexWriter.java
@@ -35,8 +35,12 @@ import org.apache.cassandra.db.DeletionTime;
 public abstract class CursorIndexWriter
 {
     protected long partitionStart;
-    // Offset within the current partition where the current index block begins.
-    protected int indexBlockStartOffset;
+    // Offset within the current partition where the current index block begins. MUST be a
+    // long, like the reference SortedTablePartitionWriter.indexBlockStartOffset: partitions
+    // can exceed 2GiB, and an int here wraps negative past Integer.MAX_VALUE — the block-cut
+    // arithmetic then cuts every row and the serialized offsets corrupt the promoted index
+    // (pinned by the giant-partition boundary run, ck_stride = rows_per_sstable).
+    protected long indexBlockStartOffset;
 
     public final void startPartition(long partitionStartPosition, long positionAfterHeader)
     {
@@ -51,14 +55,14 @@ public abstract class CursorIndexWriter
         notePosition(position);
     }
 
-    public final int indexBlockStartOffset()
+    public final long indexBlockStartOffset()
     {
         return indexBlockStartOffset;
     }
 
     protected final void notePosition(long endOfRowPosition)
     {
-        indexBlockStartOffset = (int) (endOfRowPosition - partitionStart);
+        indexBlockStartOffset = endOfRowPosition - partitionStart;
     }
 
     protected final long currentOffsetInPartition(long position)
